Jenkins: convert run script to Jenkinsfile

+++ bin/make/ipa-cal.sh
@@ -10,9 +10,12 @@ set -e
 # Command line builds alway make a fresh framework
 banner "Ensure the calabash.framework"
 
-FRAMEWORK="calabash.framework"
-rm -rf "${FRAMEWORK}"
-make framework
+
+if [ "${MAKE_FRAMEWORK}" != "0" ]; then
+  FRAMEWORK="calabash.framework"
+  rm -rf "${FRAMEWORK}"
+  make framework
+fi
 
 banner "Preparing to build LPTestTarget"
 
@@ -55,32 +58,17 @@ info "Prepared archive directory"
 
 banner "Building ${IPA}"
 
-if [ -z "${CODE_SIGN_IDENTITY}" ]; then
-  COMMAND_LINE_BUILD=1 xcrun xcodebuild \
-    -SYMROOT="${XC_BUILD_DIR}" \
-    -derivedDataPath "${XC_BUILD_DIR}" \
-    -project "${XC_PROJECT}" \
-    -scheme "${XC_TARGET}" \
-    -configuration "${XC_CONFIG}" \
-    -sdk iphoneos \
-    ARCHS="armv7 armv7s arm64" \
-    VALID_ARCHS="armv7 armv7s arm64" \
-    ONLY_ACTIVE_ARCH=NO \
-    build | $XC_PIPE
-else
-  COMMAND_LINE_BUILD=1 xcrun xcodebuild \
-    CODE_SIGN_IDENTITY="${CODE_SIGN_IDENTITY}" \
-    -SYMROOT="${XC_BUILD_DIR}" \
-    -derivedDataPath "${XC_BUILD_DIR}" \
-    -project "${XC_PROJECT}" \
-    -scheme "${XC_TARGET}" \
-    -configuration "${XC_CONFIG}" \
-    -sdk iphoneos \
-    ARCHS="armv7 armv7s arm64" \
-    VALID_ARCHS="armv7 armv7s arm64" \
-    ONLY_ACTIVE_ARCH=NO \
-    build | $XC_PIPE
-fi
+COMMAND_LINE_BUILD=1 xcrun xcodebuild \
+  -SYMROOT="${XC_BUILD_DIR}" \
+  -derivedDataPath "${XC_BUILD_DIR}" \
+  -project "${XC_PROJECT}" \
+  -scheme "${XC_TARGET}" \
+  -configuration "${XC_CONFIG}" \
+  -sdk iphoneos \
+  ARCHS="armv7 armv7s arm64" \
+  VALID_ARCHS="armv7 armv7s arm64" \
+  ONLY_ACTIVE_ARCH=NO \
+  build | $XC_PIPE
 
 EXIT_CODE=${PIPESTATUS[0]}
 
@@ -138,6 +126,14 @@ info "Copied features to ${XTC_DIR}/"
 
 ditto_or_exit cucumber/config/xtc-profiles.yml "${XTC_DIR}/cucumber.yml"
 info "Copied cucumber/config/xtc-profiles.yml to ${XTC_DIR}/"
+
+cat >"${XTC_DIR}/Gemfile" <<EOF
+source "https://rubygems.org"
+
+gem "calabash-cucumber"
+gem "rspec", "~> 3.0"
+gem "xamarin-test-cloud"
+EOF
 
 ditto_or_exit "${INSTALLED_IPA}" "${XTC_DIR}/"
 info "Copied ${IPA} to ${XTC_DIR}/"
